set xid in sendRequest
We need to ensure that outgoing requests order matches the order of the
requestsQueue so that we can detect when we receive an out of order
response given that zookeeper processes requests in a FIFO order for
clients. This means we need to set the xid of requests at the same time
we add to the queue, which in turn means we need to special case
handling of ping packets with xid=-2 whilst we use the standard write
flow for sending pings.

Also remove headerBuf on client, it could lead to races if writers do
not use the write goroutine, I may add this back soon.