wait until a byte is readable before reading packets
We can't rely on only reading responses to requests we send as Zookeeper
will send async notifications to us, eg. watches. Instead just try to
read a single byte from the bufio reader then unread it before starting
the deadline to read the full packet.

Pull setting the deadline out of readResponse so that it no longer needs
both the conn and the bufio reader, this also allows us to calcaulte
recvTimeout outside of the recv loop.